JPA repository, dependency injection, database seeding

+++ src/main/java/com/educandoweb/course/entities/User.java
@@ -15,7 +15,7 @@ public class User implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	private String name; 
 	private String email;
 	private String phone;
@@ -25,7 +25,7 @@ public class User implements Serializable {
 
 	}
 
-	public User(long id, String name, String email, String phone, String password) {
+	public User(Integer id, String name, String email, String phone, String password) {
 		this.id = id;
 		this.name = name;
 		this.email = email;
@@ -33,11 +33,11 @@ public class User implements Serializable {
 		this.password = password;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
